Rename ide-server.ts to language-client.ts
