+++ extensions/ql-vscode/src/language-support/language-client.ts
